Backend/requirements_microservice.txt
# Flask Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Document Processing (RAG Components - LangChain)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langsmith>=0.1.17,<0.2.0

# Vector Database
chromadb>=0.4.18

# Basic Utilities
numpy>=1.26.0
python-dotenv==1.0.0
pydantic>=2.7.4
requests==2.31.0
aiofiles==23.2.1

# OpenAI Client (for LLM microservice communication)
openai>=1.0.0

# Optional: Image Processing (if needed by CRM or other features)
Pillow>=10.2.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
